requirements.txt
numpy
pandas
scikit-learn
nltk
spacy
transformers
torch
tensorflow
flask
matplotlib
seaborn
pytest
hydra-core
mlflow
fastapi
uvicorn[standard]
typer